solved joining tables using normal slice
